Sell by fractions of a unit
Court time is sold by the hour and laundry by the kilo, and the previous POS
already carried lines like 2.5 hours and 3.6071 - 154 of them came over in
the import. quantity was declared Int in the schema, so those sales could not
even be read back, and the cart's own field floored anything typed into it.

Quantity is a Float now, on a sale item, a refund line and the sale input,
and the validators ask for a positive number rather than a whole one. The
cart accepts decimals and no longer rounds a price down through parseInt.

Also: tables keep a height and scroll inside it, with the header pinned. A
wide table's horizontal bar used to sit at the very bottom of the table,
below the fold on a small screen, so the columns off to the right looked
missing rather than scrollable. The bar is drawn at rest now rather than
appearing only once the table is touched.

+++ schemas/sale.schema.ts
@@ -21,12 +21,14 @@ export const saleSchema = gql`
     product: Product
     snapshotName: String
     snapshotPrice: Float
-    quantity: Int
+    # Fractional on purpose: court time is sold by the hour, laundry by the
+    # kilo, and the previous POS already carried lines like 2.5 and 3.6071.
+    quantity: Float
     discount: Float
     price: Float
     subTotal: Float
     total: Float
-    refundedQuantity: Int
+    refundedQuantity: Float
   }
 
   type SaleSettlement {
@@ -48,7 +50,7 @@ export const saleSchema = gql`
   type SaleRefundItem {
     itemIndex: Int
     snapshotName: String
-    quantity: Int
+    quantity: Float
     amount: Float
   }
 
@@ -62,7 +64,7 @@ export const saleSchema = gql`
 
   input RefundItemInput {
     itemIndex: Int!
-    quantity: Int!
+    quantity: Float!
   }
 
   type SalePayment {
@@ -159,7 +161,7 @@ export const saleSchema = gql`
     product: ID
     snapshotName: String
     snapshotPrice: Float
-    quantity: Int
+    quantity: Float
     discount: Float
     price: Float
     subTotal: Float
